feat(suite): refactoring of UpdateTooltipRow so it can be reused in onter tooltips as well
